Introduce discriminator rollback + code cleanup
Instead of just training the generator every k iteration, let the discriminator train one iteration, save it's state, train for another k-1 iterations, train the generator, and then roll back the discriminator to it's last saved state. This sort of "lookahead" allows for the generator to use discriminator predictions "ahead of time" in an attempt to make GAN training stable. highly experimental.. and slow!